feat: 게시글 작성하기 레이아웃 구현 전 공통 색 추가 및 라우팅 설정
- BoomerangColors 추가
- CommunityPosting 추가
