--- src/pages/Community/Posting.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export const Posting: React.FC = () => {
-  return <div>positng</div>;
-};
