+++ src/templates/Community/CommunityPosting.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import { BoomerangColors } from '@/utils/colors';
+import { Container } from '@chakra-ui/react';
+
+export const CommunityPosting: React.FC = () => {
+  return (
+    <Container bg={BoomerangColors.calmWhite} maxW={1024}>
+      <div>posting</div>
+    </Container>
+  );
+};
