+++ src/utils/colors.ts
@@ -2,4 +2,5 @@ export const BoomerangColors = {
   blue: '#4488FF',
   white: '#FFFFFF',
   deepBlue: '#176CFF',
+  calmWhite: '#F3F3F3',
 };
